+++ R/np.distribution.bw.R
@@ -135,7 +135,7 @@ npudistbw.NULL <-
     )
     t.names <- NULL
     if(!is.data.frame(dat) && !is.matrix(dat))
-      t.names <- deparse(substitute(dat))
+      t.names <- paste(deparse(substitute(dat)), collapse = "")
 
     dat = toFrame(dat)
     
@@ -1242,7 +1242,7 @@ npudistbw.default <-
     nomad.opts <- .np_nomad_normalize_user_opts(nomad.opts, "npudistbw")
     t.names <- NULL
     if(!is.data.frame(dat) && !is.matrix(dat))
-      t.names <- deparse(substitute(dat))
+      t.names <- paste(deparse(substitute(dat)), collapse = "")
 
     dat <- toFrame(dat)